[assembly-preparer] Refactor the loading of assemblies to a separate step.
